pyproject.toml
[project]
name = "smart-extractor"
dynamic = ["version"]
description = "自学习网页变化监控与决策辅助系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # 核心依赖
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "instructor>=1.0",
    "openai>=1.0",
    # 网页抓取
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "httpx>=0.27",
    "lxml>=5.0",
    # 容错与重试
    "tenacity>=8.0",
    # CLI 界面
    "typer>=0.12",
    "rich>=13.0",
    # 日志
    "loguru>=0.7",
    # 数据处理与存储
    "pandas>=2.0",
    # Web 界面
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    # 配置
    "pyyaml>=6.0",
    "tiktoken>=0.12.0",
    "openpyxl>=3.1",
    "python-docx>=1.1",
]

[project.scripts]
smart-extractor = "smart_extractor.cli:app"
smart-extractor-session = "smart_extractor.tools.manual_session:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/smart_extractor/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/smart_extractor"]
